[Data] Silence some conversion warnings

+++ src/components/data/choice.cpp
@@ -95,7 +95,7 @@ bool data::Choice::ChoiceAction::setup(Model& model) {
     auto& choice{static_cast<Choice&>(model)};
 
     assert(mChoice >= -1);
-    assert(mChoice < choice.mNumChoices);
+    assert(mChoice < static_cast<int32>(choice.mNumChoices));
 
     // shouldPerform is effectively the setup function for an action. It is
     // called only once and called no matter where the action comes from. Thus
@@ -103,7 +103,7 @@ bool data::Choice::ChoiceAction::setup(Model& model) {
     if (choice.mFilter) choice.mFilter(choice, mChoice);
 
     assert(mChoice >= -1);
-    assert(mChoice < choice.mNumChoices);
+    assert(mChoice < static_cast<int32>(choice.mNumChoices));
 
     return choice.mIdx != mChoice;
 }
